+++ pkg/xds/envoy/listeners/filter_chain_configurers.go
@@ -1,7 +1,6 @@
 package listeners
 
 import (
-	envoy_listener "github.com/envoyproxy/go-control-plane/envoy/config/listener/v3"
 	envoy_hcm "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
 
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
@@ -30,14 +29,6 @@ func Tracing(backend *mesh_proto.TracingBackend, service string) FilterChainBuil
 		Backend: backend,
 		Service: service,
 	})
-}
-
-func TLSInspector() ListenerBuilderOpt {
-	return AddListenerConfigurer(&v3.TLSInspectorConfigurer{})
-}
-
-func OriginalDstForwarder() ListenerBuilderOpt {
-	return AddListenerConfigurer(&v3.OriginalDstForwarderConfigurer{})
 }
 
 func StaticEndpoints(virtualHostName string, paths []*envoy_common.StaticEndpointPath) FilterChainBuilderOpt {
@@ -81,15 +72,6 @@ func SourceMatcher(address string) FilterChainBuilderOpt {
 	})
 }
 
-func InboundListener(listenerName string, address string, port uint32, protocol core_xds.SocketAddressProtocol) ListenerBuilderOpt {
-	return AddListenerConfigurer(&v3.InboundListenerConfigurer{
-		Protocol:     protocol,
-		ListenerName: listenerName,
-		Address:      address,
-		Port:         port,
-	})
-}
-
 func NetworkRBAC(statsName string, rbacEnabled bool, permission *core_mesh.TrafficPermissionResource) FilterChainBuilderOpt {
 	if !rbacEnabled {
 		return FilterChainBuilderOptFunc(nil)
@@ -99,28 +81,6 @@ func NetworkRBAC(statsName string, rbacEnabled bool, permission *core_mesh.Traff
 		StatsName:  statsName,
 		Permission: permission,
 	})
-}
-
-func OutboundListener(listenerName string, address string, port uint32, protocol core_xds.SocketAddressProtocol) ListenerBuilderOpt {
-	return AddListenerConfigurer(&v3.OutboundListenerConfigurer{
-		Protocol:     protocol,
-		ListenerName: listenerName,
-		Address:      address,
-		Port:         port,
-	})
-}
-
-func TransparentProxying(transparentProxying *mesh_proto.Dataplane_Networking_TransparentProxying) ListenerBuilderOpt {
-	virtual := transparentProxying.GetRedirectPortOutbound() != 0 && transparentProxying.GetRedirectPortInbound() != 0
-	if virtual {
-		return AddListenerConfigurer(&v3.TransparentProxyingConfigurer{})
-	}
-
-	return ListenerBuilderOptFunc(nil)
-}
-
-func NoBindToPort() ListenerBuilderOpt {
-	return AddListenerConfigurer(&v3.TransparentProxyingConfigurer{})
 }
 
 func TcpProxy(statsName string, clusters ...envoy_common.Cluster) FilterChainBuilderOpt {
@@ -227,12 +187,6 @@ func HttpOutboundRoute(service string, routes envoy_common.Routes, dpTags mesh_p
 	})
 }
 
-func FilterChain(builder *FilterChainBuilder) ListenerBuilderOpt {
-	return AddListenerConfigurer(&ListenerFilterChainConfigurerV3{
-		builder: builder,
-	})
-}
-
 func MaxConnectAttempts(retry *core_mesh.RetryResource) FilterChainBuilderOpt {
 	if retry == nil || retry.Spec.Conf.GetTcp() == nil {
 		return FilterChainBuilderOptFunc(nil)
@@ -262,35 +216,6 @@ func Timeout(timeout *mesh_proto.Timeout_Conf, protocol core_mesh.Protocol) Filt
 		Conf:     timeout,
 		Protocol: protocol,
 	})
-}
-
-func DNS(vips map[string]string, emptyDnsPort uint32) ListenerBuilderOpt {
-	return AddListenerConfigurer(&v3.DNSConfigurer{
-		VIPs:         vips,
-		EmptyDNSPort: emptyDnsPort,
-	})
-}
-
-func ConnectionBufferLimit(bytes uint32) ListenerBuilderOpt {
-	return AddListenerConfigurer(
-		v3.ListenerMustConfigureFunc(func(l *envoy_listener.Listener) {
-			l.PerConnectionBufferLimitBytes = util_proto.UInt32(bytes)
-		}))
-}
-
-func EnableReusePort(enable bool) ListenerBuilderOpt {
-	return AddListenerConfigurer(
-		v3.ListenerMustConfigureFunc(func(l *envoy_listener.Listener) {
-			// TODO(jpeach) in Envoy 1.20, this field is deprecated in favor of EnableReusePort.
-			l.ReusePort = enable
-		}))
-}
-
-func EnableFreebind(enable bool) ListenerBuilderOpt {
-	return AddListenerConfigurer(
-		v3.ListenerMustConfigureFunc(func(l *envoy_listener.Listener) {
-			l.Freebind = util_proto.Bool(enable)
-		}))
 }
 
 // ServerHeader sets the value that the HttpConnectionManager will write
